+++ multi-token-edit/scripts/picker.js
@@ -1,5 +1,6 @@
 import { SUPPORTED_PLACEABLES } from './constants.js';
 import { LinkerAPI } from './linker/linker.js';
+import { Mouse3D } from './mouse3d.js';
 import { getDataBounds, getPresetDataCenterOffset } from './presets/utils.js';
 import { pickerSelectMultiLayerDocuments } from './utils.js';
 
@@ -81,8 +82,6 @@ export class Picker {
       this._mirrorX = false;
       this._mirrorY = false;
 
-      const levelsActive = game.modules.get('levels')?.active;
-
       const centerOnCursor = () => {
         return preview.center && !(layer.name === 'TokenLayer' && preview.previewData.size === 1);
       };
@@ -90,24 +89,24 @@ export class Picker {
       // Position offset to center preview over the mouse
       let offset;
       if (centerOnCursor()) offset = getPresetDataCenterOffset(preview.previewData);
-      else offset = { x: 0, y: 0 };
+      else offset = { x: 0, y: 0, z: 0 };
+
+      if (!game.Levels3DPreview?._active) delete offset.z; // We don't want to perform z axis transform if not on 3D canvas
 
       const setPositions = function (pos) {
         if (!pos) return;
         if (centerOnCursor()) offset = getPresetDataCenterOffset(preview.previewData);
         if (preview.snap && layer && !game.keyboard.isModifierActive(KeyboardManager.MODIFIER_KEYS.SHIFT)) {
-          // v12
-          if (layer.getSnappedPoint) {
-            pos = layer.getSnappedPoint(pos);
-          } else {
-            pos = canvas.grid.getSnappedPosition(pos.x, pos.y, layer.gridPrecision);
-          }
+          const snapped = layer.getSnappedPoint(pos);
+          snapped.z = pos.z;
+          pos = snapped;
         }
 
         // calculate transform
         const pd = previews[0].document;
         const b = getDataBounds(pd.documentName, pd);
         let transform = { x: pos.x - b.x1 - offset.x, y: pos.y - b.y1 - offset.y };
+        if (pos.z != null) transform.z = pos.z - b.z1 - offset.z;
 
         // Instant transforms
         if (Picker._rotation != 0) {
@@ -130,34 +129,6 @@ export class Picker {
           const documentName = doc.documentName;
           DataTransform.apply(documentName, preview._pData ?? doc, pos, transform, preview);
 
-          // TODO: 3D Preview
-          // if (preview._l3dPreview) {
-          //   try {
-          //     preview._l3dPreview.collision = false;
-          //     const mPos = game.Levels3DPreview.interactionManager.canvas3dMousePosition;
-          //     const cPos = game.Levels3DPreview.interactionManager.camera.position;
-
-          //     const intersects = game.Levels3DPreview.interactionManager.computeSightCollisionFrom3DPositions(
-          //       cPos,
-          //       mPos,
-          //       'collision',
-          //       false,
-          //       false,
-          //       false,
-          //       true
-          //     );
-
-          //     if (intersects[0]) {
-          //       const mesh = preview._l3dPreview.mesh;
-          //       mesh.position.x = intersects[0].point.x;
-          //       mesh.position.y = intersects[0].point.y;
-          //       mesh.position.z = intersects[0].point.z;
-          //     }
-          //   } catch (e) {
-          //     console.log(e);
-          //   }
-          // }
-
           // =====
           // Hacks
           // =====
@@ -170,10 +141,13 @@ export class Picker {
             doc.sort = preview._meSort + 10000;
           }
 
-          if (doc.elevation != null) {
+          if (!game.Levels3DPreview?._active && doc.elevation != null) {
             if (!preview._meElevation) preview._meElevation = doc.elevation;
             doc.elevation = preview._meElevation + 10000;
           }
+
+          // For some reason collision bool is refreshed after creation of the preview
+          if (preview._l3dPreview) preview._l3dPreview.collision = false;
 
           // Special position update conditions
           // - Region: We need to simulate doc update via `_onUpdate` call
@@ -213,38 +187,48 @@ export class Picker {
           setPositions(event.data.getLocalPosition(pickerOverlay));
         }
       });
+
       //setTimeout(() => setPositions(canvas.mousePosition), 50);
       pickerOverlay.setPositions = setPositions;
     }
 
     if (!preview?.previewOnly) {
-      pickerOverlay.hitArea = canvas.dimensions.rect;
-      pickerOverlay.cursor = 'crosshair';
-      pickerOverlay.interactive = true;
-      pickerOverlay.zIndex = 5;
-      pickerOverlay.on('remove', () => pickerOverlay.off('pick'));
-      pickerOverlay.on('mousedown', (event) => {
-        Picker.boundStart = event.data.getLocalPosition(pickerOverlay);
-      });
-      pickerOverlay.on('mouseup', (event) => {
-        Picker.boundEnd = event.data.getLocalPosition(pickerOverlay);
-      });
-      pickerOverlay.on('click', (event) => {
-        if (event.nativeEvent.which == 2) {
-          this.callback?.(null);
-        } else {
-          const minX = Math.min(this.boundStart.x, this.boundEnd.x);
-          const maxX = Math.max(this.boundStart.x, this.boundEnd.x);
-          const minY = Math.min(this.boundStart.y, this.boundEnd.y);
-          const maxY = Math.max(this.boundStart.y, this.boundEnd.y);
-          this.callback?.({ start: { x: minX, y: minY }, end: { x: maxX, y: maxY } });
-        }
-        this.destroy();
-      });
+      if (game.Levels3DPreview?._active) {
+        Mouse3D.activate({
+          mouseMoveCallback: Picker.feedPos.bind(Picker),
+          mouseClickCallback: Picker.resolve.bind(Picker),
+          mouseWheelClickCallback: Picker.destroy.bind(Picker),
+        });
+      } else {
+        pickerOverlay.hitArea = canvas.dimensions.rect;
+        pickerOverlay.cursor = 'crosshair';
+        pickerOverlay.interactive = true;
+        pickerOverlay.zIndex = 5;
+        pickerOverlay.on('remove', () => pickerOverlay.off('pick'));
+        pickerOverlay.on('mousedown', (event) => {
+          Picker.boundStart = event.data.getLocalPosition(pickerOverlay);
+        });
+        pickerOverlay.on('mouseup', (event) => {
+          Picker.boundEnd = event.data.getLocalPosition(pickerOverlay);
+        });
+        pickerOverlay.on('click', (event) => {
+          if (event.nativeEvent.which == 2) {
+            this.callback?.(null);
+          } else {
+            const minX = Math.min(this.boundStart.x, this.boundEnd.x);
+            const maxX = Math.max(this.boundStart.x, this.boundEnd.x);
+            const minY = Math.min(this.boundStart.y, this.boundEnd.y);
+            const maxY = Math.max(this.boundStart.y, this.boundEnd.y);
+            this.callback?.({ start: { x: minX, y: minY }, end: { x: maxX, y: maxY } });
+          }
+          this.destroy();
+        });
 
-      canvas.stage.addChild(pickerOverlay);
+        canvas.stage.addChild(pickerOverlay);
+      }
     }
     this.pickerOverlay = pickerOverlay;
+    setTimeout(() => this.feedPos(canvas.mousePosition), 100);
   }
 
   static feedPos(pos) {
@@ -266,12 +250,11 @@ export class Picker {
         this.pickerOverlay.previewDocuments.forEach((name) => {
           const layer = canvas.getLayerByEmbeddedName(name);
           if (layer) {
-            // TODO: 3D Preview
-            // if (game.Levels3DPreview?._active) {
-            //   layer.preview.children.forEach((c) => {
-            //     c._l3dPreview?.destroy();
-            //   });
-            // }
+            if (game.Levels3DPreview?._active) {
+              layer.preview.children.forEach((c) => {
+                c._l3dPreview?.destroy();
+              });
+            }
             layer.clearPreviewContainer();
           }
         });
@@ -281,6 +264,7 @@ export class Picker {
       this.pickerOverlay.children?.forEach((c) => c.destroy(true));
       this.callback?.(null);
       this.pickerOverlay = null;
+      Mouse3D.deactivate();
     }
     this.callback = null;
   }
@@ -310,14 +294,27 @@ export class Picker {
     // lets simplify by overriding this function to make sure the preview is always visible
     Picker._overridePlaceableVisibility(object);
 
-    // TODO: 3D Preview
-    // if (game.Levels3DPreview._active) {
-    //   if (documentName === 'Tile') {
-    //     game.Levels3DPreview.createTile(object);
-    //     object._l3dPreview = game.Levels3DPreview.tiles[object.id];
-    //     console.log(object._l3dPreview);
-    //   }
-    // }
+    // 3D Canvas
+    if (game.Levels3DPreview?._active) {
+      if (documentName === 'Tile') {
+        game.Levels3DPreview.createTile(object);
+        const l3dPreview = game.Levels3DPreview.tiles[object.id];
+
+        l3dPreview.castShadow = false;
+        l3dPreview.collision = false;
+
+        object._l3dPreview = l3dPreview;
+      } else if (documentName === 'Token') {
+        // Tokens get async loaded without a way to await them
+        // We'll need to retrieve the 3D token when the transforms are actually getting applied
+        game.Levels3DPreview.addToken(object);
+        object._l3dPreview = null;
+      } else if (documentName === 'AmbientLight') {
+        game.Levels3DPreview.addLight(object);
+        const l3dPreview = game.Levels3DPreview.lights[object.id];
+        object._l3dPreview = l3dPreview;
+      }
+    }
 
     return object;
   }
@@ -859,13 +856,22 @@ export class DataTransform {
       doc.config.dim = data.config.dim;
       doc.config.bright = data.config.bright;
       if (data.elevation) doc.elevation = data.elevation;
+
+      if (preview._l3dPreview) {
+        const pos = game.Levels3DPreview.ruler.constructor.posCanvasTo3d({
+          x: doc.x,
+          y: doc.y,
+          z: doc.elevation,
+        });
+        const mesh = preview._l3dPreview.mesh;
+        mesh.position.x = pos.x;
+        mesh.position.y = pos.y;
+        mesh.position.z = pos.z;
+      }
     }
   }
 
   static transformTile(data, origin, transform, preview) {
-    // 3D support
-    const depth = data.flags?.['levels-3d-preview']?.depth;
-
     if (transform.scale != null) {
       const scale = transform.scale;
       data.x *= scale;
@@ -874,7 +880,8 @@ export class DataTransform {
       data.height *= scale;
 
       // 3D Support
-      if (depth != null && depth != '') data.flags['levels-3d-preview'].depth *= scale;
+      const depth = data.flags?.['levels-3d-preview']?.depth;
+      if (depth != null && depth != '') data.flags['levels-3d-preview'].depth = depth * scale;
       if (data.elevation != null) {
         data.elevation *= scale;
       }
@@ -917,7 +924,26 @@ export class DataTransform {
       doc.rotation = data.rotation;
       doc.texture.scaleX = data.texture.scaleX;
       doc.texture.scaleY = data.texture.scaleY;
-      if (data.elevation) doc.elevation = data.elevation;
+      if (data.elevation != null) doc.elevation = data.elevation;
+
+      if (preview._l3dPreview && preview._l3dPreview.mesh) {
+        const pos = game.Levels3DPreview.ruler.constructor.posCanvasTo3d({
+          x: doc.x + doc.width / 2,
+          y: doc.y + doc.height / 2,
+          z: doc.elevation,
+        });
+        const mesh = preview._l3dPreview.mesh;
+        mesh.position.x = pos.x;
+        mesh.position.y = pos.y;
+        mesh.position.z = pos.z;
+
+        if (transform.scale != null) {
+          mesh.scale.multiplyScalar(transform.scale);
+        }
+        if (transform.rotation != null) {
+          mesh.rotation.y += game.Levels3DPreview.THREE.MathUtils.degToRad(-transform.rotation);
+        }
+      }
     }
   }
 
@@ -1009,29 +1035,29 @@ export class DataTransform {
     if (transform.rotation != null) {
       const dr = Math.toRadians(transform.rotation % 360);
       let rectCenter = {
-        x: data.x + (data.width * (grid.sizeX ?? grid.w)) / 2,
-        y: data.y + (data.height * (grid.sizeY ?? grid.h)) / 2,
+        x: data.x + (data.width * grid.sizeX) / 2,
+        y: data.y + (data.height * grid.sizeY) / 2,
       };
       [rectCenter.x, rectCenter.y] = this.rotatePoint(origin.x, origin.y, rectCenter.x, rectCenter.y, dr);
-      data.x = rectCenter.x - (data.width * (grid.sizeX ?? grid.w)) / 2;
-      data.y = rectCenter.y - (data.height * (grid.sizeY ?? grid.h)) / 2;
+      data.x = rectCenter.x - (data.width * grid.sizeX) / 2;
+      data.y = rectCenter.y - (data.height * grid.sizeY) / 2;
       data.rotation = (data.rotation + Math.toDegrees(dr)) % 360;
     }
 
     if (transform.mirrorX || transform.mirrorY) {
       let rectCenter = {
-        x: data.x + (data.width * (grid.sizeX ?? grid.w)) / 2,
-        y: data.y + (data.height * (grid.sizeY ?? grid.h)) / 2,
+        x: data.x + (data.width * grid.sizeX) / 2,
+        y: data.y + (data.height * grid.sizeY) / 2,
       };
       if (transform.mirrorX) {
         rectCenter.x = origin.x - (rectCenter.x - origin.x);
         data.texture.scaleX *= -1;
-        data.x = rectCenter.x - (data.width * (grid.sizeX ?? grid.w)) / 2;
+        data.x = rectCenter.x - (data.width * grid.sizeX) / 2;
       }
       if (transform.mirrorY) {
         rectCenter.y = origin.y - (rectCenter.y - origin.y);
         data.texture.scaleY *= -1;
-        data.y = rectCenter.y - (data.height * (grid.sizeY ?? grid.h)) / 2;
+        data.y = rectCenter.y - (data.height * grid.sizeY) / 2;
       }
       data.rotation = 180 - (data.rotation - 180);
     }
@@ -1047,6 +1073,20 @@ export class DataTransform {
       doc.texture.scaleX = data.texture.scaleX;
       doc.texture.scaleY = data.texture.scaleY;
       if (data.elevation) doc.elevation = data.elevation;
+      if (preview.hasOwnProperty('_l3dPreview')) {
+        preview._l3dPreview = game.Levels3DPreview.tokens[preview.id];
+        if (!preview._l3dPreview) return;
+
+        const pos = game.Levels3DPreview.ruler.constructor.posCanvasTo3d({
+          x: doc.x + (doc.width * canvas.grid.sizeX) / 2,
+          y: doc.y + (doc.height * canvas.grid.sizeY) / 2,
+          z: doc.elevation,
+        });
+        const mesh = preview._l3dPreview.mesh;
+        mesh.position.x = pos.x;
+        mesh.position.y = pos.y;
+        mesh.position.z = pos.z;
+      }
     }
   }
 
@@ -1106,8 +1146,8 @@ export async function editPreviewPlaceables() {
         foundry.utils.setProperty(data, 'flags.token-attacher.prototypeAttached', prototypeAttached);
         foundry.utils.setProperty(data, 'flags.token-attacher.grid', {
           size: canvas.grid.size,
-          w: canvas.grid.sizeX ?? canvas.grid.w, // v12
-          h: canvas.grid.sizeY ?? canvas.grid.h, // v12
+          w: canvas.grid.sizeX,
+          h: canvas.grid.sizeY,
         });
       }
     }
@@ -1137,7 +1177,6 @@ export async function editPreviewPlaceables() {
           const diff = foundry.utils.diffObject(originalData[i], data[i]);
           if (!foundry.utils.isEmpty(diff)) {
             diff._id = originalData[i]._id;
-            delete diff.flags;
             updates.push(diff);
           }
         }